@@ -1,81 +1,23 @@
 {
   "name": "frenchies-spots-webui",
-  "version": "1.0.0",
-  "main": "index.js",
-  "engines": {
-    "node": "18.16.1"
-  },
+  "version": "0.1.0",
+  "private": true,
   "scripts": {
-    "start": "expo start --clear",
-    "android": "expo start --clear --android",
-    "ios": "expo start --clear --ios",
-    "web": "expo start --clear --web",
-    "eject": "expo eject",
-    "dev": "expo start --clear --web",
-    "build": "npx expo export:web"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
-    "@apollo/client": "^3.7.10",
-    "@babel/preset-env": "7.1.6",
-    "@expo/match-media": "^0.4.0",
-    "@expo/ngrok": "4.1.0",
-    "@hookform/resolvers": "^3.0.0",
-    "@mapbox/search-js-react": "1.0.0-beta.14",
-    "@react-native-async-storage/async-storage": "~1.17.3",
-    "@react-native-material/core": "^1.3.7",
-    "@react-native-picker/picker": "2.4.8",
-    "@react-navigation/native": "^6.1.6",
-    "@react-navigation/native-stack": "^6.9.12",
-    "@rnmapbox/maps": "10.0.0-rc.5",
-    "core-js": "^3.29.1",
-    "expo": "~47.0.0",
-    "expo-cli": "6.3.2",
-    "expo-constants": "~14.0.2",
-    "expo-font": "~11.0.1",
-    "expo-image-picker": "~14.0.2",
-    "expo-keep-awake": "~11.0.1",
-    "expo-linking": "~3.3.1",
-    "expo-location": "~15.0.1",
-    "expo-modules-autolinking": "~1.0.0",
-    "expo-router": "^0.0.40",
-    "expo-screen-orientation": "~5.0.1",
-    "expo-status-bar": "~1.4.4",
-    "i18next": "^22.4.13",
-    "lodash": "^4.17.21",
-    "mapbox-gl": "2.0.0",
-    "metro": "~0.73.3",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
-    "react-hook-form": "^7.43.8",
-    "react-i18next": "^12.2.0",
-    "react-map-gl": "^7.0.21",
-    "react-map-gl-geocoder": "2.2.0",
-    "react-native": "0.70.8",
-    "react-native-gesture-handler": "~2.8.0",
-    "react-native-get-random-values": "~1.8.0",
-    "react-native-paper": "^5.4.1",
-    "react-native-reanimated": "^3.4.2",
-    "react-native-root-siblings": "^4.1.1",
-    "react-native-root-toast": "^3.3.1",
-    "react-native-safe-area-context": "4.4.1",
-    "react-native-screens": "~3.18.2",
-    "react-native-svg": "13.4.0",
-    "react-native-swiper-flatlist": "^3.0.18",
-    "react-native-vector-icons": "^9.2.0",
-    "react-native-web": "~0.18.10",
-    "react-native-webview": "11.23.1",
-    "react-responsive": "^9.0.2",
-    "uuid": "^7.0.3",
-    "yup": "^1.0.2"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@expo/webpack-config": "^18.0.1",
-    "@types/lodash": "^4.14.191",
-    "@types/react": "^18.0.22",
-    "@types/react-native": "^0.69.15",
-    "@types/webpack": "^5.28.0",
-    "typescript": "^4.7.4"
-  },
-  "private": true
+    "@types/node": "20.5.6",
+    "@types/react": "18.2.21",
+    "@types/react-dom": "18.2.7",
+    "eslint": "8.48.0",
+    "eslint-config-next": "13.4.19",
+    "next": "13.4.19",
+    "next-pwa": "^5.6.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "typescript": "5.2.2"
+  }
 }
